requirements.txt
pymupdf
Flask
flask_cors
pinecone
protobuf
python-dotenv
Werkzeug
google-genai